requirements.txt
python-dotenv
requests
json
halo